+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Fixed workflow for publishing docker images on tagged releases.
 - Fixed delete samples bug.
 - Fixed `upload_sample.py` add to group bug.
+- Fixed close basket bug.
+- Fixed find closest samples bug.
 
 ### Changed